--- .env
@@ -1,7 +1,0 @@
-# Environment variables declared in this file are automatically made available to Prisma.
-# See the documentation for more detail: https://pris.ly/d/prisma-schema#accessing-environment-variables-from-the-schema
-
-# Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB and CockroachDB.
-# See the documentation for all the connection string options: https://pris.ly/d/connection-strings
-
-DATABASE_URL="prisma+postgres://accelerate.prisma-data.net/?api_key=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhcGlfa2V5IjoiMDFKWEo0NjY2VzY5WjRaQVhQTUZNQzNBTTAiLCJ0ZW5hbnRfaWQiOiJkNTM0M2FmNDQxZDVlYTI0ZTdkNTUxN2IzMjJmN2MwOGE0ZjkyOWRhNDlkODg2ZDA0ZmExMzEzZGNmYWNjMGZiIiwiaW50ZXJuYWxfc2VjcmV0IjoiYTBjZDJmZGQtMmI4Ni00MTM2LWEwNDYtNmU5MjI5M2IxNmU0In0.W2y25ttDd5RWppGInwtlCOk4E2887apHha4zzzHsAUM"
