+++ package.json
@@ -11,21 +11,14 @@
     "format": "npm run format --workspaces",
     "dev:web": "npm run dev --workspace=@sose2025/web",
     "build:web": "npm run build --workspace=@sose2025/web",
-    "start:web": "npm run start --workspace=@sose2025/web"
-  },
-  "prisma": {
-    "seed": "tsx prisma/seed.ts"
+    "start:web": "npm run start --workspace=@sose2025/web",
+    "prisma:generate": "npm run prisma:generate --workspace=@sose2025/web",
+    "prisma:migrate": "npm run prisma:migrate --workspace=@sose2025/web",
+    "prisma:seed": "npm run prisma:seed --workspace=@sose2025/web",
+    "prepare": "npm run prepare --workspace=@sose2025/web"
+
   },
   "workspaces": [
     "packages/*"
-  ],
-  "devDependencies": {
-    "dotenv-cli": "^8.0.0",
-    "prisma": "^6.9.0",
-    "tsx": "^4.20.2"
-  },
-  "dependencies": {
-    "@prisma/client": "^6.9.0",
-    "@prisma/extension-accelerate": "^2.0.1"
-  }
+  ]
 }
