+++ packages/web/.env.example
@@ -1,0 +1,5 @@
+# Change API-KEY to your actual API key from Prisma Data Platform when using it.
+# DATABASE_URL="prisma+postgres://accelerate.prisma-data.net/?api_key=API-KEY"
+
+# For local database development, change dummy credentials to your actual database credentials.
+DATABASE_URL="postgresql://ducks_user:ducks_secret@localhost:5432/ducks_db"
