+++ packages/web/.gitignore
@@ -31,7 +31,10 @@ yarn-error.log*
 .pnpm-debug.log*
 
 # env files (can opt-in for committing if needed)
-# .env*
+.env*
+
+# except example env
+!.env.example
 
 # vercel
 .vercel
