--- prisma/data.ts
@@ -1,329 +1,0 @@
-export const rubberDuckData = [
-    {
-        id: 1,
-        name: "Classic Duck",
-        price: 999,
-        color: "yellow",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A classic yellow rubber duck",
-    },
-    {
-        id: 2,
-        name: "Red Duck",
-        price: 1199,
-        color: "red",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A red rubber duck",
-    },
-    {
-        id: 3,
-        name: "Green Duck",
-        price: 1299,
-        color: "green",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A green rubber duck",
-    },
-    {
-        id: 4,
-        name: "Blue Duck",
-        price: 1399,
-        color: "blue",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A blue rubber duck",
-    },
-    {
-        id: 5,
-        name: "Large Duck",
-        price: 1849,
-        color: "yellow",
-        material: "TPE",
-        size: "xl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 1.5,
-        description: "A large yellow rubber duck",
-    },
-    {
-        id: 6,
-        name: "Small Duck",
-        price: 799,
-        color: "yellow",
-        material: "PVC",
-        size: "s",
-        brand: "Tiny Toys",
-        origin: "Germany",
-        producer: "Tiny Toys Inc.",
-        weight: 0.3,
-        description: "A small yellow rubber duck",
-    },
-    {
-        id: 7,
-        name: "Extra Large Duck",
-        price: 2599,
-        color: "yellow",
-        material: "TPE",
-        size: "xxl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 2.5,
-        description: "An extra large yellow rubber duck",
-    },
-    {
-        id: 8,
-        name: "Small Red Duck",
-        price: 899,
-        color: "red",
-        material: "PVC",
-        size: "s",
-        brand: "Tiny Toys",
-        origin: "Germany",
-        producer: "Tiny Toys Inc.",
-        weight: 0.3,
-        description: "A small red rubber duck",
-    },
-    {
-        id: 9,
-        name: "Large Green Duck",
-        price: 1999,
-        color: "green",
-        material: "TPE",
-        size: "xl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 1.5,
-        description: "A large green rubber duck",
-    },
-    {
-        id: 10,
-        name: "Blue Duck with Hat",
-        price: 1449,
-        color: "blue",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A blue rubber duck with a hat",
-    },
-    {
-        id: 11,
-        name: "Yellow Duck with Sunglasses",
-        price: 1299,
-        color: "yellow",
-        material: "PVC",
-        size: "s",
-        brand: "Best Toys",
-        origin: "France",
-        producer: "Best Toys Inc.",
-        weight: 0.4,
-        description: "A yellow rubber duck with sunglasses",
-    },
-    {
-        id: 12,
-        name: "Giant Green Duck",
-        price: 2999,
-        color: "green",
-        material: "PP",
-        size: "xxl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 3.0,
-        description: "A giant green rubber duck",
-    },
-    {
-        id: 13,
-        name: "Hoodie Duck",
-        price: 1599,
-        color: "yellow",
-        material: "PS",
-        size: "l",
-        brand: "Cool Toys",
-        origin: "Germany",
-        producer: "Cool Toys Inc.",
-        weight: 0.8,
-        description: "A yellow rubber duck with a hoodie",
-    },
-    {
-        id: 14,
-        name: "King Duck",
-        price: 1499,
-        color: "red",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A red rubber duck with a crown",
-    },
-    {
-        id: 15,
-        name: "Heart Duck",
-        price: 1399,
-        color: "red",
-        material: "PS",
-        size: "m",
-        brand: "Best Toys",
-        origin: "France",
-        producer: "Best Toys Inc.",
-        weight: 0.6,
-        description: "A red rubber duck with a heart design in the chest",
-    },
-    {
-        id: 16,
-        name: "Pirate Duck",
-        price: 999,
-        color: "yellow",
-        material: "PVC",
-        size: "s",
-        brand: "Cool Toys",
-        origin: "Germany",
-        producer: "Cool Toys Inc.",
-        weight: 0.4,
-        description: "A yellow rubber duck dressed as a pirate",
-    },
-    {
-        id: 17,
-        name: "Keychain Duck",
-        price: 499,
-        color: "blue",
-        material: "PVC",
-        size: "s",
-        brand: "Tiny Toys",
-        origin: "Germany",
-        producer: "Tiny Toys Inc.",
-        weight: 0.2,
-        description: "A small blue rubber duck keychain",
-    },
-    {
-        id: 18,
-        name: "Screaming Duck",
-        price: 1099,
-        color: "green",
-        material: "ABS",
-        size: "l",
-        brand: "Cool Toys",
-        origin: "Germany",
-        producer: "Cool Toys Inc.",
-        weight: 0.7,
-        description: "A green rubber duck that screams when squeezed",
-    },
-    {
-        id: 19,
-        name: "Giant Glowing Duck",
-        price: 3599,
-        color: "blue",
-        material: "TPE",
-        size: "xxl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 4.0,
-        description: "A giant blue rubber duck that glows in the dark",
-    },
-    {
-        id: 20,
-        name: "Unicorn Duck",
-        price: 1749,
-        color: "yellow",
-        material: "PVC",
-        size: "xl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 2.2,
-        description: "A big yellow rubber duck with a unicorn horn",
-    },
-    {
-        id: 21,
-        name: "Hippy Duck",
-        price: 1499,
-        color: "green",
-        material: "PP",
-        size: "m",
-        brand: "Best Toys",
-        origin: "France",
-        producer: "Best Toys Inc.",
-        weight: 0.5,
-        description: "A green rubber duck with a hippy design",
-    },
-    {
-        id: 22,
-        name: "Flying Pirate Programming Duck with Hat on a Broom",
-        price: 19999,
-        color: "blue",
-        material: "Carbon Fiber",
-        size: "xxl",
-        brand: "Special Toys",
-        origin: "Gurkistan",
-        producer: "Special Toys Inc.",
-        weight: 15.5,
-        description:
-            "A blue rubber duck dressed as a pirate, flying on a broom with a hat programming in Python. It is a special edition duck for programmers.",
-    },
-    {
-        id: 23,
-        name: "Fly Duck",
-        price: 545,
-        color: "green",
-        material: "PS",
-        size: "s",
-        brand: "Tiny Toys",
-        origin: "Germany",
-        producer: "Tiny Toys Inc.",
-        weight: 0.3,
-        description: "A small green rubber duck that can fly",
-    },
-    {
-        id: 24,
-        name: "Duck with Wings",
-        price: 1299,
-        color: "yellow",
-        material: "PVC",
-        size: "m",
-        brand: "Classic Toys",
-        origin: "USA",
-        producer: "Classic Toys Inc.",
-        weight: 0.5,
-        description: "A yellow rubber duck with wings",
-    },
-    {
-        id: 25,
-        name: "Drinking Duck",
-        price: 4599,
-        color: "red",
-        material: "PVC",
-        size: "xl",
-        brand: "Giant Toys",
-        origin: "China",
-        producer: "Giant Toys Co.",
-        weight: 2.5,
-        description: "A red rubber duck that can drinking Vodka (Bottle included)",
-    },
-    // Add more ducks if needed
-];
